PathwayEditor execute the function of DataChanged and SelectChanged when the object is selected in PathwayEditor. Because PathwayEditor execute DataChanged after SelectChanged, there are no selected object on EntityListWindow and ObjectList.
git-svn-id: file:///home/moriyoshi/chaperone.e-cell.org/var/opt/subversion/ecell-ide/trunk@622 fe1fbaf7-ad0a-4b3d-a902-8cb9c273ac6b

+++ ide/ObjectList/ObjectList/ObjectList.cs
@@ -840,6 +840,7 @@ namespace EcellLib.ObjectList
                !m_currentModelID.Equals(modelID))
                 return;
 
+            if (key == data.key) return;
             ObjectListOfType olot = m_dict[type];
             olot.DeleteObjectWithFull(key);
 
@@ -1100,4 +1101,39 @@ namespace EcellLib.ObjectList
         }
         #endregion
     }
+
+    public class CustomComparer : IComparer
+    {
+        private int sortOrder;
+        private Comparer comparer;
+
+        public CustomComparer(SortOrder order)
+        {
+            this.sortOrder = (order == SortOrder.Descending ? -1 : 1);
+            this.comparer = new Comparer(
+                System.Globalization.CultureInfo.CurrentCulture);
+        }
+
+        //並び替え方を定義する
+        public int Compare(object x, object y)
+        {
+            int result = 0;
+
+            DataGridViewRow rowx = (DataGridViewRow)x;
+            DataGridViewRow rowy = (DataGridViewRow)y;
+
+            //はじめの列のセルの値を比較し、同じならば次の列を比較する
+            for (int i = 0; i < rowx.Cells.Count; i++)
+            {
+                result = this.comparer.Compare(
+                    rowx.Cells[i].Value, rowy.Cells[i].Value);
+                if (result != 0)
+                    break;
+            }
+
+            //結果を返す
+            return result * this.sortOrder;
+        }
+    }
+
 }
